LRQA-29755 Get test results from batch builds rather than all build types
